minor verification fix

+++ src/bindings/python/nixl_utils.cpp
@@ -38,7 +38,7 @@ void ba_buf(uintptr_t addr, int size) {
 }
 
 //JUST FOR TESTING
-int verify_n(uintptr_t addr1, int num, int size) {
+int verify_n(uintptr_t addr1, uint8_t num, int size) {
     for(int i = 0; i<size; i++) {
 	    if(((uint8_t*) addr1)[i] != num) return 0;
     }
